+++ src/queryselector/mod.rs
@@ -1,5 +1,10 @@
+/// Query selector iterator
 pub mod iter;
+/// Query selector iterable
+pub mod iterable;
+/// Query selector parser
 pub mod parser;
+/// Query selector
 pub mod selector;
 
 pub use iter::*;