+++ src/util.rs
@@ -1,4 +1,4 @@
-#[inline]
+#[inline(never)]
 pub fn is_ident(c: u8) -> bool {
     (b'0'..=b'9').contains(&c)
         || (b'A'..=b'Z').contains(&c)
@@ -7,28 +7,30 @@ pub fn is_ident(c: u8) -> bool {
         || c == b'_'
 }
 
-#[inline]
-pub fn is_strict_whitespace(c: u8) -> bool {
-    c == b' '
-}
-
-#[inline]
+#[inline(never)]
 pub fn find_slow(haystack: &[u8], needle: u8) -> Option<usize> {
     haystack.iter().position(|&c| c == needle)
 }
 
-#[inline]
+#[inline(never)]
 pub fn find_multi_slow<const N: usize>(haystack: &[u8], needle: [u8; N]) -> Option<usize> {
     haystack.iter().position(|c| needle.contains(c))
 }
 
 #[cfg(feature = "simd")]
 pub fn find_fast(haystack: &[u8], needle: u8) -> Option<usize> {
-    use std::simd::*;
+    use std::{ptr, simd::*};
+
+    #[inline(never)]
+    #[cold]
+    fn unlikely_find(haystack: &[u8], needle: u8) -> Option<usize> {
+        find_slow(haystack, needle)
+    }
 
     let len = haystack.len();
+    let ptr = haystack.as_ptr();
     if len < 16 {
-        return find_slow(haystack, needle);
+        return unlikely_find(haystack, needle);
     }
 
     let mut i = 0;
@@ -36,9 +38,7 @@ pub fn find_fast(haystack: &[u8], needle: u8) -> Option<usize> {
 
     while i <= len - 16 {
         let mut bytes = [0; 16];
-        for j in 0..16 {
-            bytes[j] = unsafe { *haystack.get_unchecked(i + j) };
-        }
+        unsafe { ptr::copy_nonoverlapping(ptr.add(i), bytes.as_mut_ptr(), 16) };
 
         let bytes = u8x16::from_array(bytes);
         let eq = bytes.lanes_eq(needle16).to_int();
@@ -55,11 +55,18 @@ pub fn find_fast(haystack: &[u8], needle: u8) -> Option<usize> {
 
 #[cfg(feature = "simd")]
 pub fn find_fast_4(haystack: &[u8], needle: [u8; 4]) -> Option<usize> {
-    use std::simd::*;
+    use std::{ptr, simd::*};
+
+    #[inline(never)]
+    #[cold]
+    fn unlikely_find(haystack: &[u8], needle: [u8; 4]) -> Option<usize> {
+        find_multi_slow(haystack, needle)
+    }
 
     let len = haystack.len();
+    let ptr = haystack.as_ptr();
     if len < 16 {
-        return find_multi_slow(haystack, needle);
+        return unlikely_find(haystack, needle);
     }
 
     let mut i = 0;
@@ -70,9 +77,7 @@ pub fn find_fast_4(haystack: &[u8], needle: [u8; 4]) -> Option<usize> {
 
     while i <= len - 16 {
         let mut bytes = [0; 16];
-        for j in 0..16 {
-            bytes[j] = unsafe { *haystack.get_unchecked(i + j) };
-        }
+        unsafe { ptr::copy_nonoverlapping(ptr.add(i), bytes.as_mut_ptr(), 16) };
 
         let bytes = u8x16::from_array(bytes);
 
@@ -98,11 +103,18 @@ pub fn search_non_ident_slow(haystack: &[u8]) -> Option<usize> {
 
 #[cfg(feature = "simd")]
 pub fn search_non_ident_fast(haystack: &[u8]) -> Option<usize> {
-    use std::simd::*;
+    use std::{ptr, simd::*};
+
+    #[inline(never)]
+    #[cold]
+    fn unlikely_search(haystack: &[u8]) -> Option<usize> {
+        search_non_ident_slow(haystack)
+    }
 
     let len = haystack.len();
+    let ptr = haystack.as_ptr();
     if len < 16 {
-        return search_non_ident_slow(haystack);
+        return unlikely_search(haystack);
     }
 
     let mut i = 0;
@@ -117,9 +129,7 @@ pub fn search_non_ident_fast(haystack: &[u8]) -> Option<usize> {
 
     while i <= len - 16 {
         let mut bytes = [0; 16];
-        for j in 0..16 {
-            bytes[j] = unsafe { *haystack.get_unchecked(i + j) };
-        }
+        unsafe { ptr::copy_nonoverlapping(ptr.add(i), bytes.as_mut_ptr(), 16) };
 
         let bytes = u8x16::from_array(bytes);
 
@@ -150,4 +160,44 @@ pub fn search_non_ident_fast(haystack: &[u8]) -> Option<usize> {
     }
 
     search_non_ident_slow(&haystack[i..]).map(|x| i + x)
+}
+
+#[inline]
+pub fn is_closing(needle: u8) -> bool {
+    #[cfg(feature = "simd")]
+    {
+        use std::simd::u8x4;
+        let sc = u8x4::from_array([b'/', b'>', 0, 0]);
+        let needle = u8x4::splat(needle);
+        let eq = needle.lanes_eq(sc);
+        eq.any()
+    }
+    #[cfg(not(feature = "simd"))]
+    {
+        let eq1 = needle == b'/';
+        let eq2 = needle == b'>';
+
+        eq1 | eq2
+    }
+}
+
+#[inline(always)]
+fn to_lower(byte: u8) -> u8 {
+    let is_upper = (byte >= b'A' && byte <= b'Z') as u8;
+    let lower = is_upper * 0x20;
+    byte + lower
+}
+
+pub fn matches_case_insensitive<const N: usize>(haystack: &[u8], needle: [u8; N]) -> bool {
+    if haystack.len() != N {
+        return false;
+    }
+
+    // LLVM seems to already generate pretty good SIMD even without explicit use
+
+    let mut mask = true;
+    for i in 0..N {
+        mask &= to_lower(haystack[i]) == needle[i];
+    }
+    mask
 }